feat: elevate channels to its own sidebar tab
Move channels configuration from a nested tab inside Settings to a
dedicated top-level sidebar item with its own route (/app/channels).
Old /app/settings/channels URLs redirect to the new location.

Fixes #514
