+++ frontend/src/App.tsx
@@ -7,6 +7,7 @@ import MemoryPage from '@/pages/MemoryPage';
 import SettingsPage from '@/pages/SettingsPage';
 import ChatPage from '@/pages/ChatPage';
 import ChecklistPage from '@/pages/ChecklistPage';
+import ChannelsPage from '@/pages/ChannelsPage';
 import { useAuth } from '@/contexts/AuthContext';
 import {
   getLoginPageElement,
@@ -43,6 +44,7 @@ export default function App() {
         <Route path="conversations/:sessionId" element={<ConversationsPage />} />
         <Route path="memory" element={<MemoryPage />} />
         <Route path="checklist" element={<ChecklistPage />} />
+        <Route path="channels" element={<ChannelsPage />} />
         <Route path="settings/:tab" element={<SettingsPage />} />
         <Route path="settings" element={<Navigate to={`/app/settings/${getDefaultSettingsTab(isPremium)}`} replace />} />
       </Route>
