+++ frontend/src/pages/SettingsPage.tsx
@@ -1,5 +1,5 @@
 import { useState, useCallback, useEffect } from 'react';
-import { useOutletContext, useParams, useNavigate } from 'react-router-dom';
+import { Navigate, useOutletContext, useParams, useNavigate } from 'react-router-dom';
 import Card from '@/components/ui/card';
 import Input from '@/components/ui/input';
 import Textarea from '@/components/ui/textarea';
@@ -8,7 +8,7 @@ import Select from '@/components/ui/select';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
 import { toast } from 'sonner';
 import api from '@/api';
-import type { ChannelConfig, ContractorProfileUpdate, ToolConfigEntry } from '@/types';
+import type { ContractorProfileUpdate, ToolConfigEntry } from '@/types';
 import type { AppShellContext } from '@/layouts/AppShell';
 import {
   getExtraSettingsTabs,
@@ -20,6 +20,11 @@ export default function SettingsPage() {
   const { tab } = useParams<{ tab: string }>();
   const navigate = useNavigate();
   const { profile, reloadProfile, isPremium, isAdmin } = useOutletContext<AppShellContext>();
+
+  // Redirect old /app/settings/channels to /app/channels
+  if (tab === 'channels') {
+    return <Navigate to="/app/channels" replace />;
+  }
 
   const extraTabs = getExtraSettingsTabs(isPremium, isAdmin);
   const activeTab = tab || 'profile';
@@ -41,7 +46,6 @@ export default function SettingsPage() {
                 <TabsTrigger value="profile">Profile</TabsTrigger>
                 <TabsTrigger value="assistant">Assistant</TabsTrigger>
                 <TabsTrigger value="heartbeat">Heartbeat</TabsTrigger>
-                <TabsTrigger value="channels">Channels</TabsTrigger>
                 <TabsTrigger value="tools">Tools</TabsTrigger>
               </>
             )}
@@ -65,7 +69,6 @@ export default function SettingsPage() {
           <TabsTrigger value="profile">Profile</TabsTrigger>
           <TabsTrigger value="assistant">Assistant</TabsTrigger>
           <TabsTrigger value="heartbeat">Heartbeat</TabsTrigger>
-          <TabsTrigger value="channels">Channels</TabsTrigger>
           <TabsTrigger value="tools">Tools</TabsTrigger>
           {extraTabs.map((t) => (
             <TabsTrigger key={t.key} value={t.key}>{t.label}</TabsTrigger>
@@ -82,10 +85,6 @@ export default function SettingsPage() {
 
         <TabsContent value="heartbeat">
           {profile && <HeartbeatTab profile={profile} onSaved={reloadProfile} />}
-        </TabsContent>
-
-        <TabsContent value="channels">
-          {profile && <ChannelsTab profile={profile} />}
         </TabsContent>
 
         <TabsContent value="tools">
@@ -382,123 +381,6 @@ function HeartbeatTab({
             {saving ? 'Saving...' : 'Save Heartbeat Settings'}
           </Button>
         </div>
-      </div>
-    </Card>
-  );
-}
-
-// --- Channels Tab ---
-
-function ChannelsTab({
-  profile,
-}: {
-  profile: { channel_identifier: string; preferred_channel: string };
-}) {
-  const connected = !!profile.channel_identifier;
-  const [config, setConfig] = useState<ChannelConfig | null>(null);
-  const [botToken, setBotToken] = useState('');
-  const [allowedUsernames, setAllowedUsernames] = useState('');
-  const [saving, setSaving] = useState(false);
-
-  useEffect(() => {
-    api.getChannelConfig().then((cfg) => {
-      setConfig(cfg);
-      setAllowedUsernames(cfg.telegram_allowed_usernames);
-    }).catch(() => {
-      // ignore fetch errors on mount
-    });
-  }, []);
-
-  const handleSave = useCallback(async () => {
-    setSaving(true);
-    try {
-      const body: Record<string, string> = {};
-      if (botToken) body.telegram_bot_token = botToken;
-      if (config && allowedUsernames !== config.telegram_allowed_usernames) {
-        body.telegram_allowed_usernames = allowedUsernames;
-      }
-      if (Object.keys(body).length === 0) {
-        toast.error('No changes to save');
-        setSaving(false);
-        return;
-      }
-      const updated = await api.updateChannelConfig(body);
-      setConfig(updated);
-      setBotToken('');
-      toast.success('Channel config updated');
-    } catch (e) {
-      toast.error((e as Error).message);
-    } finally {
-      setSaving(false);
-    }
-  }, [botToken, allowedUsernames, config]);
-
-  return (
-    <Card>
-      <div className="grid gap-4">
-        <Field label="Bot Token">
-          {config === null ? (
-            <p className="text-sm text-muted-foreground">Loading...</p>
-          ) : (
-            <>
-              <div className="mb-2">
-                {config.telegram_bot_token_set ? (
-                  <span className="inline-flex items-center gap-1.5 text-sm">
-                    <span className="w-2 h-2 rounded-full bg-green-500 inline-block" />
-                    Configured
-                  </span>
-                ) : (
-                  <span className="inline-flex items-center gap-1.5 text-sm">
-                    <span className="w-2 h-2 rounded-full bg-red-500 inline-block" />
-                    Not configured
-                  </span>
-                )}
-              </div>
-              <Input
-                type="password"
-                value={botToken}
-                onChange={(e) => setBotToken(e.target.value)}
-                placeholder={config.telegram_bot_token_set ? 'Enter new token to replace' : 'Paste bot token from @BotFather'}
-              />
-            </>
-          )}
-        </Field>
-        <Field label="Allowed Usernames">
-          <Input
-            value={allowedUsernames}
-            onChange={(e) => setAllowedUsernames(e.target.value)}
-            placeholder='Comma-separated @usernames, or * for all'
-          />
-          <p className="text-xs text-muted-foreground mt-1">
-            Controls which Telegram users can message your bot.
-          </p>
-        </Field>
-        <div className="flex justify-end">
-          <Button onClick={handleSave} disabled={saving || config === null}>
-            {saving ? 'Saving...' : 'Save Channel Config'}
-          </Button>
-        </div>
-        <hr className="border-border" />
-        <Field label="User Connection">
-          {connected ? (
-            <div className="flex items-center gap-2">
-              <span className="inline-flex items-center gap-1.5 text-sm">
-                <span className="w-2 h-2 rounded-full bg-green-500 inline-block" />
-                Connected
-              </span>
-              <span className="text-xs text-muted-foreground">
-                Chat ID: {profile.channel_identifier}
-              </span>
-            </div>
-          ) : (
-            <p className="text-sm text-muted-foreground">
-              Send a message to your bot on Telegram to connect.
-            </p>
-          )}
-        </Field>
-        <Field label="Active Channel">
-          <p className="text-sm">{profile.preferred_channel || 'webchat'}</p>
-        </Field>
       </div>
     </Card>
   );
